Migrate mssql extracts to assessments/common/ and pin MSSQL_SCHEMAS
Move the inlined mssql_* schema definitions out of the back-compat
save_resultset_to_db wrapper (where they were placed in #2151) into
mssql/common/schemas.py as MSSQL_SCHEMAS. Switch the 9 call sites in
mssql/info_extract.py and 4 call sites in mssql/activity_extract.py to
save_to_duckdb, looking the schema up explicitly via
MSSQL_SCHEMAS[f"mssql_{table_name}"]. Also update both helper and
arguments_loader imports to the new shared location at
assessments/common/.

This removes the cross-cutting import chain (mssql/* -> synapse/common/*)
that #2151 had to introduce because there was no shared helper module to
import from. With assessments/common/ in place, both backends own only
their backend-specific things (queries, schemas, connectors) and share
the DuckDB I/O via the new top-level common package.

Part of #2093.
