+++ src/databricks/labs/lakebridge/resources/assessments/mssql/common/schemas.py
@@ -1,0 +1,25 @@
+"""Pinned per-table DuckDB schemas for MSSQL profiler extracts.
+
+See ``synapse/common/schemas.py`` for the rationale; the same dtype-drift
+concern applies whenever an extract appends across runs.
+"""
+
+from __future__ import annotations
+
+MSSQL_SCHEMAS: dict[str, str] = {
+    # Activity extract
+    "mssql_query_stats": "DBID VARCHAR, CREATION_TIME TIMESTAMP, LAST_EXECUTION_TIME TIMESTAMP, EXECUTION_COUNT BIGINT, TOTAL_WORKER_TIME BIGINT, TOTAL_ELAPSED_TIME BIGINT, TOTAL_ROWS BIGINT, COMMAND VARCHAR, COMMAND_TYPE VARCHAR, EXTRACT_TS TIMESTAMP",
+    "mssql_proc_stats": "DATABASE_ID BIGINT, DB_NAME VARCHAR, OBJECT_ID BIGINT, OBJECT_NAME VARCHAR, TYPE VARCHAR, LAST_EXECUTION_TIME TIMESTAMP, EXECUTION_COUNT BIGINT, TOTAL_WORKER_TIME BIGINT, TOTAL_ELAPSED_TIME BIGINT, EXTRACT_TS TIMESTAMP",
+    "mssql_sessions": "SESSION_ID BIGINT, LOGIN_TIME TIMESTAMP, PROGRAM_NAME VARCHAR, CLIENT_INTERFACE_NAME VARCHAR, LOGIN_NAME VARCHAR, STATUS VARCHAR, CPU_TIME BIGINT, MEMORY_USAGE BIGINT, TOTAL_SCHEDULED_TIME BIGINT, TOTAL_ELAPSED_TIME BIGINT, LAST_REQUEST_START_TIME TIMESTAMP, LAST_REQUEST_END_TIME TIMESTAMP, IS_USER_PROCESS BOOLEAN, ROW_COUNT BIGINT, DATABASE_ID BIGINT, DB_NAME VARCHAR, EXTRACT_TS TIMESTAMP",
+    "mssql_cpu_utilization": "RECORD_ID BIGINT, EVENTTIME TIMESTAMP, SYSTEMIDLE BIGINT, SQLPROCESSUTILIZATION BIGINT, EXTRACT_TS TIMESTAMP",
+    # Info extract
+    "mssql_sys_info": "CPU_TICKS BIGINT, MS_TICKS BIGINT, CPU_COUNT BIGINT, HYPERTHREAD_RATIO BIGINT, PHYSICAL_MEMORY_KB BIGINT, VIRTUAL_MEMORY_KB BIGINT, COMMITTED_KB BIGINT, COMMITTED_TARGET_KB BIGINT, VISIBLE_TARGET_KB BIGINT, STACK_SIZE_IN_BYTES BIGINT, OS_QUANTUM BIGINT, OS_ERROR_MODE BIGINT, OS_PRIORITY_CLASS BIGINT, MAX_WORKERS_COUNT BIGINT, SCHEDULER_COUNT BIGINT, SCHEDULER_TOTAL_COUNT BIGINT, DEADLOCK_MONITOR_SERIAL_NUMBER BIGINT, SQLSERVER_START_TIME_MS_TICKS BIGINT, SQLSERVER_START_TIME TIMESTAMP, AFFINITY_TYPE BIGINT, AFFINITY_TYPE_DESC VARCHAR, PROCESS_KERNEL_TIME_MS BIGINT, PROCESS_USER_TIME_MS BIGINT, TIME_SOURCE BIGINT, TIME_SOURCE_DESC VARCHAR, VIRTUAL_MACHINE_TYPE BIGINT, VIRTUAL_MACHINE_TYPE_DESC VARCHAR, SOFTNUMA_CONFIGURATION BIGINT, SOFTNUMA_CONFIGURATION_DESC VARCHAR, PROCESS_PHYSICAL_AFFINITY VARCHAR, SQL_MEMORY_MODEL BIGINT, SQL_MEMORY_MODEL_DESC VARCHAR, SOCKET_COUNT BIGINT, CORES_PER_SOCKET BIGINT, NUMA_NODE_COUNT BIGINT, CONTAINER_TYPE BIGINT, CONTAINER_TYPE_DESC VARCHAR, EXTRACT_TS TIMESTAMP",
+    "mssql_databases": "DB_ID VARCHAR, NAME VARCHAR, COLLATION_NAME VARCHAR, CREATE_DATE TIMESTAMP, EXTRACT_TS TIMESTAMP",
+    "mssql_tables": "TABLE_CATALOG VARCHAR, TABLE_SCHEMA VARCHAR, TABLE_NAME VARCHAR, TABLE_TYPE VARCHAR",
+    "mssql_views": "TABLE_CATALOG VARCHAR, TABLE_SCHEMA VARCHAR, TABLE_NAME VARCHAR, CHECK_OPTION VARCHAR, IS_UPDATABLE VARCHAR, VIEW_DEFINITION VARCHAR",
+    "mssql_columns": "TABLE_CATALOG VARCHAR, TABLE_SCHEMA VARCHAR, TABLE_NAME VARCHAR, COLUMN_NAME VARCHAR, ORDINAL_POSITION BIGINT, COLUMN_DEFAULT VARCHAR, IS_NULLABLE VARCHAR, DATA_TYPE VARCHAR, CHARACTER_MAXIMUM_LENGTH DOUBLE, CHARACTER_OCTET_LENGTH DOUBLE, NUMERIC_PRECISION DOUBLE, NUMERIC_PRECISION_RADIX DOUBLE, NUMERIC_SCALE DOUBLE, DATETIME_PRECISION DOUBLE, CHARACTER_SET_CATALOG VARCHAR, CHARACTER_SET_SCHEMA VARCHAR, CHARACTER_SET_NAME VARCHAR, COLLATION_CATALOG VARCHAR, COLLATION_SCHEMA VARCHAR, COLLATION_NAME VARCHAR, DOMAIN_CATALOG VARCHAR, DOMAIN_SCHEMA VARCHAR, DOMAIN_NAME VARCHAR",
+    "mssql_indexed_views": "INDEXED_VIEW_NAME VARCHAR, SCHEMA_NAME VARCHAR, INDEX_NAME VARCHAR, INDEX_TYPE VARCHAR, INDEX_ID VARCHAR, EXTRACT_TS VARCHAR",
+    "mssql_routines": "CREATED TIMESTAMP, DATA_TYPE VARCHAR, IS_DETERMINISTIC VARCHAR, IS_IMPLICITLY_INVOCABLE VARCHAR, IS_NULL_CALL VARCHAR, IS_USER_DEFINED_CAST VARCHAR, LAST_ALTERED TIMESTAMP, MAX_DYNAMIC_RESULT_SETS BIGINT, NUMERIC_PRECISION DOUBLE, NUMERIC_PRECISION_RADIX DOUBLE, NUMERIC_SCALE DOUBLE, ROUTINE_BODY VARCHAR, ROUTINE_CATALOG VARCHAR, ROUTINE_DEFINITION VARCHAR, ROUTINE_NAME VARCHAR, ROUTINE_SCHEMA VARCHAR, ROUTINE_TYPE VARCHAR, SCHEMA_LEVEL_ROUTINE VARCHAR, SPECIFIC_CATALOG VARCHAR, SPECIFIC_NAME VARCHAR, SPECIFIC_SCHEMA VARCHAR, SQL_DATA_ACCESS VARCHAR",
+    "mssql_db_sizes": "DATABASE_NAME VARCHAR, FILENAME VARCHAR, TYPE_DESC VARCHAR, CURRENTSIZEMB VARCHAR, FREESPACEINMB VARCHAR, MAXSIZE BIGINT, EXTRACT_TS TIMESTAMP",
+    "mssql_table_sizes": "TABLENAME VARCHAR, ROWCOUNT BIGINT, RESERVEDMB BIGINT, USEDMB BIGINT, UNUSEDMB BIGINT, DATAMB BIGINT, INDEXMB BIGINT, EXTRACT_TS TIMESTAMP",
+}
