+++ src/databricks/labs/lakebridge/resources/assessments/mssql/info_extract.py
@@ -5,10 +5,11 @@ from databricks.labs.blueprint.entrypoint import get_logger
 
 from databricks.labs.lakebridge.connections.credential_manager import create_credential_manager
 from databricks.labs.lakebridge.assessments import PRODUCT_NAME
+from databricks.labs.lakebridge.resources.assessments.common.cli import arguments_loader
+from databricks.labs.lakebridge.resources.assessments.common.duckdb_helpers import save_to_duckdb
 from databricks.labs.lakebridge.resources.assessments.mssql.common.connector import get_sqlserver_reader
 from databricks.labs.lakebridge.resources.assessments.mssql.common.queries import MSSQLQueries
-from databricks.labs.lakebridge.resources.assessments.synapse.common.duckdb_helpers import save_resultset_to_db
-from databricks.labs.lakebridge.resources.assessments.synapse.common.functions import arguments_loader
+from databricks.labs.lakebridge.resources.assessments.mssql.common.schemas import MSSQL_SCHEMAS
 
 logger = get_logger(__file__)
 
@@ -39,63 +40,63 @@ def execute():
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         print(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Databases
         table_name = "databases"
         table_query = MSSQLQueries.get_databases()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Tables
         table_name = "tables"
         table_query = MSSQLQueries.get_tables()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Views
         table_name = "views"
         table_query = MSSQLQueries.get_views()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Columns
         table_name = "columns"
         table_query = MSSQLQueries.get_columns()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Indexed views
         table_name = "indexed_views"
         table_query = MSSQLQueries.get_indexed_views()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Routines
         table_name = "routines"
         table_query = MSSQLQueries.get_routines()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Database sizes
         table_name = "db_sizes"
         table_query = MSSQLQueries.get_db_sizes()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         # Table sizes
         table_name = "table_sizes"
         table_query = MSSQLQueries.get_table_sizes()
         logger.info(f"Loading '{table_name}' for SQL server: {server_name}")
         result = connection.fetch(table_query)
-        save_resultset_to_db(result, f"mssql_{table_name}", db_path, mode=mode)
+        save_to_duckdb(result.to_df(), f"mssql_{table_name}", db_path, mode=mode, schema=MSSQL_SCHEMAS[f"mssql_{table_name}"])
 
         print(json.dumps({"status": "success", "message": "All data loaded successfully loaded successfully"}))
